Add optional catch all route
