+++ app/[team]/[[...section]]/layout.tsx
