--- app/[team]/page.tsx
@@ -1,7 +1,0 @@
-export default function Page() {
-  return (
-    <section className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-      <p className="mt-8">This is the home dashboard page content.</p>
-    </section>
-  );
-}
